refactor out of react hook
