+++ packages/code-block/src/lib/formatter/formatter.ts
@@ -1,24 +1,13 @@
+import type { Editor } from '@udecode/plate';
+
+import type { TCodeBlockElement } from '../types';
+
 import { formatJson, isValidJson } from './jsonFormatter';
 
 const supportedLanguages = new Set(['json']);
 
 export const isLangSupported = (lang?: string): boolean =>
   Boolean(lang && supportedLanguages.has(lang));
-
-export const formatCode = (code: string, lang?: string): string => {
-  if (!isLangSupported(lang)) {
-    return '';
-  }
-
-  switch (lang) {
-    case 'json': {
-      return formatJson(code);
-    }
-    default: {
-      return code;
-    }
-  }
-};
 
 export const isValidSyntax = (code: string, lang?: string): boolean => {
   if (!isLangSupported(lang)) {
@@ -31,6 +20,39 @@ export const isValidSyntax = (code: string, lang?: string): boolean => {
     }
     default: {
       return false;
+    }
+  }
+};
+
+export const formatCodeBlock = (
+  editor: Editor,
+  {
+    element,
+  }: {
+    element: TCodeBlockElement;
+  }
+) => {
+  const { lang } = element;
+
+  if (!lang || !isLangSupported(lang)) {
+    return;
+  }
+
+  const code = editor.api.string(element);
+
+  if (isValidSyntax(code, lang)) {
+    const formattedCode = formatCode(code, lang);
+    editor.tf.insertText(formattedCode, { at: element });
+  }
+};
+
+const formatCode = (code: string, lang?: string): string => {
+  switch (lang) {
+    case 'json': {
+      return formatJson(code);
+    }
+    default: {
+      return code;
     }
   }
 };
