+++ packages/code-block/src/lib/formatter/jsonFormatter.spec.tsx
@@ -1,21 +1,21 @@
-import { formatCode, isValidSyntax } from './formatter';
+import { formatJson, isValidJson } from './jsonFormatter';
 
 describe('JsonFormatter', () => {
   it('should detect valid JSON', () => {
     const json = '{ "name": "ChatGPT", "type": "AI" }';
-    const isValid = isValidSyntax(json, 'json');
+    const isValid = isValidJson(json);
     expect(isValid).toBe(true);
   });
 
   it('should detect invalid JSON', () => {
     const json = '{ name: "ChatGPT", type: AI }';
-    const isValid = isValidSyntax(json, 'json');
+    const isValid = isValidJson(json);
     expect(isValid).toBe(false);
   });
 
   it('should format JSON', () => {
     const json = '{"name":"ChatGPT","type":"AI"}';
-    const formattedJson = formatCode(json, 'json');
+    const formattedJson = formatJson(json);
     const expected = `{
   "name": "ChatGPT",
   "type": "AI"
@@ -25,7 +25,7 @@ describe('JsonFormatter', () => {
 
   it('should not format invalid JSON', () => {
     const json = '{ name: "ChatGPT", type: AI }';
-    const formattedJson = formatCode(json, 'json');
+    const formattedJson = formatJson(json);
     expect(formattedJson).toBe(json);
   });
 });
