+++ packages/code-block/src/react/hooks/index.ts
@@ -1,9 +1,6 @@
-/**
- * @file Automatically generated by barrelsby.
- */
+/** @file Automatically generated by barrelsby. */
 
 export * from './useCodeBlockCombobox';
 export * from './useCodeBlockElement';
-export * from './useCodeBlockFormat';
 export * from './useCodeSyntaxLeaf';
 export * from './useToggleCodeBlockButton';
